+++ src/vipersci/carto/tri2gpkg.py
@@ -89,7 +89,7 @@ def arg_parser():
     parser.add_argument(
         "--s_srs",
         default="+proj=cart +a=1737400 +b=1737400",
-        help="The source CRS or SRS of the .tri10 data.",
+        help="The source CRS or SRS of the .tri10 data. Default: %(default)s",
     )
     parser.add_argument(
         "--t_srs",
@@ -103,7 +103,7 @@ def arg_parser():
         help="This text will be used as the title of the data value field in "
         "the output file.  If there are commas, this text is split on "
         "the commas and whitespace stripped, to produce a list of value "
-        "fields in the output.",
+        "fields in the output. Default: %(default)s",
     )
     parser.add_argument(
         "--value_columns",
@@ -111,7 +111,8 @@ def arg_parser():
         help="This should be a comma-separated list of integers (or just one) "
         "specifying which columns from the .tri file get the names from "
         "-v.  The first 9 columns (zero is the first) have facet "
-        "coordinates, so typically this value (or list) starts at 9.",
+        "coordinates, so typically this value (or list) starts at 9. "
+        "Default: %(default)s",
     )
     parser.add_argument(
         "--value_file",
@@ -285,7 +286,7 @@ def replace_with(replacement_val, replacement_check, value):
     if float(value) == replacement_check:
         return replacement_val
     else:
-        return value
+        return float(value)
 
 
 # def get_wkt_value(transformer, tokens: list):
@@ -314,7 +315,3 @@ def replace_with(replacement_val, replacement_check, value):
 #         ]
 #     )
 #     return (f"POLYGON Z (({c_list}))")
-
-
-if __name__ == "__main__":
-    main()