+++ src/vipersci/vis/pds/create_raw.py
@@ -1,7 +1,14 @@
 """Creates Raw VIS PDS Products.
 
-For now, this program has a variety of pre-set data,
+This module builds "raw" VIS data products from input data.  A "raw" VIS
+data product consists of a TIFF file and a JSON file containing meta-data.
+The meta data can optionally be inserted into a database, and/or used to
+generate a PDS4 XML label.
+
+For now, this program still has a variety of hard-coded elements,
 that will eventually be extracted from telemetry.
+
+The command-line version is primarily to aide testing.
 """
 
 # Copyright 2022, United States Government as represented by the
@@ -33,10 +40,8 @@ from importlib import resources
 import io
 import json
 import logging
-import sys
 from typing import Union
 from pathlib import Path
-from warnings import warn
 
 from genshi.template import MarkupTemplate
 import numpy as np
@@ -53,6 +58,8 @@ from vipersci import util
 
 logger = logging.getLogger(__name__)
 
+ImageType = Union[npt.NDArray[np.uint16], npt.NDArray[np.uint8]]
+
 
 def arg_parser():
     parser = argparse.ArgumentParser(
@@ -66,17 +73,17 @@ def arg_parser():
         "postgresql://postgres:NotTheDefault@localhost/visdb",
     )
     parser.add_argument(
-        "--json",
-        type=Path,
-        required=True,
-        help="JSON file, with raw product meta-data.",
+        "--nojson",
+        action="store_false",
+        dest="json",
+        help="Disables creation of .json output, which is the default.",
     )
     parser.add_argument(
         "-t",
         "--template",
         type=Path,
         help="Genshi XML file template.  Will default to the raw-template.xml "
-        "file distributed with the module.",
+        "file distributed with the module.  Only relevant when --xml is provided.",
     )
     parser.add_argument(
         "--image",
@@ -88,17 +95,22 @@ def arg_parser():
     parser.add_argument(
         "--tiff",
         type=Path,
-        help="Optional pre-existing TIFF file.  This file will be inspected "
-        "and an .xml label will be written.  If --image is given, this will "
-        "be ignored.",
+        help="Optional pre-existing TIFF file (presumably created by this program). "
+        "This file will be inspected and its information added to the output. "
+        "If --image is given, this option will be ignored.",
     )
     parser.add_argument(
         "-o",
         "--output_dir",
         type=Path,
         default=Path.cwd(),
-        help="Output directory for label. Defaults to current working directory.",
+        help="Output directory for label. Defaults to current working directory.  "
+        "Output file names are fixed based on product_id, and will be over-written.",
     )
+    parser.add_argument(
+        "-x", "--xml", action="store_true", help="Create a PDS4 .XML label file."
+    )
+    parser.add_argument("input", type=Path, help="JSON file containing metadata.")
     return parser
 
 
@@ -106,14 +118,14 @@ def main():
     args = arg_parser().parse_args()
     util.set_logger(args.verbose)
 
-    with open(args.json) as f:
+    with open(args.input) as f:
         jsondata = json.load(f)
 
     # I'm not sure where these are coming from, let's hard-code them for now:
     metadata = {
         "mission_phase": "TEST",
-        "purpose": "Navigation",
-        "onboard_compression_ratio": 64,
+        # "purpose": "Engineering",
+        # "onboard_compression_ratio": 64,
     }
 
     # This allows values in jsondata to override the hard-coded values above.
@@ -127,11 +139,21 @@ def main():
         image = None
 
     if args.dburl is None:
-        create(metadata, image, args.output_dir, None, args.template)
+        create(
+            metadata, image, args.output_dir, None, args.json, args.xml, args.template
+        )
     else:
         engine = create_engine(args.dburl)
         session_maker = sessionmaker(engine, future=True)
-        create(metadata, image, args.output_dir, session_maker, args.template)
+        create(
+            metadata,
+            image,
+            args.output_dir,
+            session_maker,
+            args.json,
+            args.xml,
+            args.template,
+        )
 
     return
 
@@ -141,8 +163,8 @@ class Creator:
     This object can be instantiated with an output directory, *outdir*, and optional
     *session* and *template_path* directories, which the object maintains.
 
-    This object can simply be called which results in a raw product being created,
-    written to disk and possibly added to the database.
+    This object can simply be called which results in a raw product TIFF file and JSON
+    file being created, written to disk and possibly added to the database.
 
     This is basically a persistent version of the create() function, so that a
     database connection can be kept alive (during a Yamcs subscription, for example),
@@ -150,25 +172,22 @@ class Creator:
 
     All of the arguments to initialize the object are optional:  If *outdir* is not
     given, the current working directory will be used (beware!).  If *session* is not
-    given, no writes to a database will occur.  If *template_path* is not given, a
-    default from this library will be used.
+    given, no writes to a database will occur.
     """
 
     def __init__(
         self,
         outdir: Path = Path.cwd(),
         session: Session = None,
-        template_path: Path = None,
     ):
         self.outdir = outdir
         self.session = session
-        self.tmpl_path = template_path
 
-    def __call__(self, metadata: dict, image: npt.NDArray[np.uint16] = None):
+    def __call__(self, metadata: dict, image: ImageType = None):
         rp = make_raw_product(metadata, image, self.outdir)
         logger.info(f"{rp.product_id} created.")
 
-        write_xml(rp.label_dict(), self.outdir, self.tmpl_path)
+        write_json(rp.asdict(), self.outdir)
 
         if self.session is not None:
             with self.session.begin() as s:
@@ -179,7 +198,7 @@ class Creator:
     def from_yamcs_parameters(self, data):
         for parameter in data.parameters:
             logger.info(f"{parameter.generation_time} - {parameter.name}")
-            # These are hard-coded until I figure out where they come from.
+            # These are hard-coded until we figure out where they come from.
             d = {
                 "bad_pixel_table_id": 0,
                 "hazlight_aft_port_on": False,
@@ -192,8 +211,6 @@ class Creator:
                 "navlight_right_on": False,
                 "mission_phase": "TEST",
                 "purpose": "Navigation",
-                "onboard_compression_ratio": 64,
-                "onboard_compression_type": parameter.name[-4:].upper(),
             }
             d.update(parameter.eng_value["imageHeader"])
             d["yamcs_name"] = parameter.name
@@ -207,9 +224,11 @@ class Creator:
 
 def create(
     metadata: dict,
-    image: Union[npt.NDArray[np.uint16], Path] = None,
+    image: Union[ImageType, Path] = None,
     outdir: Path = Path.cwd(),
-    session: Session = None,
+    session: Union[Session, sessionmaker] = None,
+    json: bool = True,
+    xml: bool = False,
     template_path: Path = None,
 ):
     """
@@ -237,9 +256,17 @@ def create(
     The *template_path* argument is passed to the write_xml() function, please see
     its documentation for details.
     """
+    # This arrangement of creating the output files first, ensures a clean
+    # db insert (otherwise the outfile-writing would need to happen in the
+    # context of the session).  However, if the db insert fails, the files
+    # already exist on disk.  Is that a problem?  Maybe that's fine?
     rp = make_raw_product(metadata, image, outdir)
 
-    write_xml(rp.label_dict(), outdir, template_path)
+    if json:
+        write_json(rp.asdict(), outdir)
+
+    if xml:
+        write_xml(rp.label_dict(), outdir, template_path)
 
     if session is not None:
         with session.begin() as s:
@@ -248,9 +275,47 @@ def create(
     return
 
 
+def check_bit_depth(pid: pds.VISID, bit_depth: Union[int, str, np.dtype]):
+    """If the provided *bit_depth* is incompatible with *pid* a ValueError will be
+    raised.
+
+    The *bit_depth* will attempt to be converted to a valid integer value the bit
+    depth of the PDS TIFF file.  The *bit_depth* argument can be an integer, a
+    numpy dtype, and even a string compatible with the PDS4
+    Array_2D_Image/Element_Array/data_type attribute.
+    """
+
+    bd = None
+    if isinstance(bit_depth, int):
+        bd = bit_depth
+    elif isinstance(bit_depth, str):
+        bd = int(bit_depth[-1]) * 8
+    elif isinstance(bit_depth, np.dtype):
+        if bit_depth == np.uint16:
+            bd = 16
+        elif bit_depth == np.uint8:
+            bd = 8
+
+    if bd is None:
+        raise ValueError(f"Cannot determine a bit-depth from {bit_depth}")
+
+    if pid.compression == "s":
+        if bd != 8:
+            raise ValueError(
+                f"This is a SLoG product ({pid}), but this image is not 8-bit, "
+                f"it is {bit_depth}"
+            )
+    else:
+        if bd != 16:
+            raise ValueError(
+                f"This product ({pid}) should be 16-bit, but it is {bit_depth}"
+            )
+    return
+
+
 def make_raw_product(
     metadata: dict,
-    image: Union[npt.NDArray[np.uint16], Path] = None,
+    image: Union[ImageType, Path] = None,
     outdir: Path = Path.cwd(),
 ) -> RawProduct:
     """
@@ -260,12 +325,13 @@ def make_raw_product(
     working directory).
     """
     rp = RawProduct(**metadata)
+    pid = pds.VISID(rp.product_id)
 
     if image is not None:
         if isinstance(image, Path):
             tif_d = tif_info(image)
         else:
-            tif_d = tif_info(write_tiff(pds.VISID(rp.product_id), image, outdir))
+            tif_d = tif_info(write_tiff(pid, image, outdir))
 
         for k in ("lines", "samples"):
             if getattr(rp, k) != tif_d[k]:
@@ -274,10 +340,11 @@ def make_raw_product(
                     f"match the value from the metadata ({getattr(rp, k)})"
                 )
 
+        check_bit_depth(pid, tif_d["file_data_type"])
+
         rp.update(tif_d)
     else:
-        rp.update({"byte_offset": None, "data_type": None})
-    rp.update(version_info(rp.product_id))
+        rp.update({"file_byte_offset": None, "file_data_type": None})
 
     rp.update(
         {
@@ -306,31 +373,22 @@ def tif_info(p: Path) -> dict:
     info = read_tiff(str(p))
     tags = info["ifds"][0]["tags"]
 
-    if info["bigEndian"]:
-        end = "MSB"
-    else:
-        end = "LSB"
-
-    if tags[258]["data"][0] != 16:
-        raise ValueError(
-            f"TIFF file has {tags[258]['data'][0]} BitsPerSample " f"expecting 16."
-        )
-    else:
-        dtype = f"Unsigned{end}2"
+    end = "MSB" if info["bigEndian"] else "LSB"
 
     d = {
-        "file_path": p.name,
+        "file_byte_offset": tags[273]["data"][0],  # Tag 273 is StripOffsets
         "file_creation_datetime": dt,
-        "md5_checksum": md5.hexdigest(),
-        "byte_offset": tags[273]["data"][0],  # Tag 273 is StripOffsets
+        # Tag 258 is bits per pixel:
+        "file_data_type": f"Unsigned{end}{int(tags[258]['data'][0] / 8)}",
+        "file_md5_checksum": md5.hexdigest(),
+        "file_path": p.name,
         "lines": tags[257]["data"][0],  # Tag 257 is ImageWidth,
         "samples": tags[256]["data"][0],  # Tag 256 is ImageWidth,
-        "data_type": dtype,
     }
     return d
 
 
-def version_info(pid: pds.VISID):
+def version_info():
     # This should reach into a database and do something smart to figure
     # out how to populate this, but for now, hardcoding:
     d = {
@@ -346,20 +404,27 @@ def version_info(pid: pds.VISID):
     return d
 
 
-def write_tiff(
-    pid: pds.VISID, image: npt.NDArray[np.uint16], outdir: Path = Path.cwd()
-) -> Path:
+def write_json(product: dict, outdir: Path = Path.cwd()):
+    """
+    Convenience function to write *product* as a JSON file in
+    *outdir*.
+    """
+    out_path = (outdir / product["product_id"]).with_suffix(".json")
+    with out_path.open("w") as f:
+        json.dump(product, f, indent=2, sort_keys=True)
+
+
+def write_tiff(pid: pds.VISID, image: ImageType, outdir: Path = Path.cwd()) -> Path:
     """
     Returns the path where a TIFF with a name based on *pid* and the array
     *image* was written in *outdir* (defaults to current working directory).
     """
-    if image.dtype != np.uint16:
-        # raise ValueError(
-        warn(f"The input image is not a uint16, it is {image.dtype}")
+
+    check_bit_depth(pid, image.dtype)
 
     desc = f"VIPER {pds.vis_instruments[pid.instrument]} {pid}"
 
-    logger.info(desc)
+    logger.debug(desc)
     outpath = (outdir / str(pid)).with_suffix(".tif")
 
     imsave(str(outpath), image, check_contrast=False, description=desc, metadata=None)
@@ -382,11 +447,10 @@ def write_xml(product: dict, outdir: Path = Path.cwd(), template_path: Path = No
     else:
         tmpl = MarkupTemplate(template_path.read_text())
 
-    stream = tmpl.generate(**product)
+    d = version_info()
+    d.update(product)
+
+    stream = tmpl.generate(**d)
     out_path = (outdir / product["product_id"]).with_suffix(".xml")
     out_path.write_text(stream.render())
     return
-
-
-if __name__ == "__main__":
-    sys.exit(main())  # pragma: no cover