+++ tests/vis/pds/test_create_raw.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-"""This module has tests for the pds.create_raw functions."""
+"""This module has tests for the vis.pds.create_raw functions."""
 
 # Copyright 2022, vipersci developers.
 #
@@ -7,105 +7,133 @@
 # The AUTHORS file and the LICENSE file are at the
 # top level of this library.
 
-# import contextlib
-# import shutil
-import datetime
+from datetime import datetime, timezone
+from pathlib import Path
 import unittest
+from unittest.mock import create_autospec, mock_open, patch
 
-# from pathlib import Path
+import numpy as np
 
-from vipersci.pds import pid
+from vipersci.pds import pid as pds
+from vipersci.vis.db.raw_products import RawProduct
+from vipersci.vis.pds import create_raw as cw
 
 
-class TestInfo(unittest.TestCase):
-    def test_tiff_info(self):
-        tuples = (
-            (
-                ("220117", "010101", "ncl"),
-                (datetime.date(2022, 1, 17), datetime.time(1, 1, 1), "ncl"),
-            ),
-            (
-                ("220117", "010101001", "aim"),
-                (datetime.datetime(2022, 1, 17, 1, 1, 1, 1000), "aim"),
-            ),
-            (
-                ("231225", "182000", "acr"),
-                (datetime.date(2023, 12, 25), datetime.time(18, 20, 0), "acr"),
-            ),
-            (
-                ("240330", "121212", "hap"),
-                (datetime.datetime(2024, 3, 30, 12, 12, 12), "hap"),
-            ),
-        )
-        for truth, t in tuples:
-            with self.subTest():
-                p = pid.VIPERID(*t)
-                self.assertTupleEqual(truth, (p.date, p.time, p.instrument))
+class TestBitDepth(unittest.TestCase):
+    def test_check_bit_depth(self):
+        pid = pds.VISID("231125-143859-ncl-a")
+        self.assertIsNone(cw.check_bit_depth(pid, 16))
+        self.assertIsNone(cw.check_bit_depth(pid, "UnsignedLSB2"))
+        self.assertIsNone(cw.check_bit_depth(pid, np.uint16(5).dtype))
 
-    def test_init_string(self):
-        string_tuples = (
-            ("220117-010101-ncl", "220117-010101-ncl"),
-            ("220117-010101-ncl", "220117-010101-NCL"),
-            ("220117-010101-ncl", "20220117-010101-ncl"),
-            ("231225-182000-acr", "231225-182000-acr ignored"),
-        )
-        for s in string_tuples:
-            with self.subTest():
-                p = pid.VIPERID(s[1])
-                self.assertEqual(s[0], str(p))
+        self.assertRaises(ValueError, cw.check_bit_depth, pid, None)
+        self.assertRaises(ValueError, cw.check_bit_depth, pid, 8)
 
-    def test_init_bad_tuples(self):
-        tuples = (
-            ("not a datetime", "ncl"),
-            ("not a date", datetime.time(1, 1, 1), "ncl"),
-            (datetime.date(2024, 1, 1), "not a time", "ncl"),
-            (datetime.datetime(2022, 1, 17, 1, 1, 1), "not an instrument"),
-        )
-        for t in tuples:
-            with self.subTest(t):
-                self.assertRaises(ValueError, pid.VIPERID, *t)
+        pids = pds.VISID("231125-143859-ncl-s")
+        self.assertIsNone(cw.check_bit_depth(pids, np.uint8(5).dtype))
+        self.assertRaises(ValueError, cw.check_bit_depth, pids, 16)
 
-    def test_init_bad_strings(self):
-        strings = (
-            "220117-010101-aaa",
-            "220117-10101-ncl",
-            "foobar",
-        )
-        for s in strings:
-            with self.subTest(s):
-                self.assertRaises(ValueError, pid.VIPERID, s)
 
-    def test_init_too_many_args(self):
+class TestMakeRaw(unittest.TestCase):
+    def setUp(self) -> None:
+        self.startUTC = datetime(2022, 1, 27, 0, 0, 0, tzinfo=timezone.utc)
+        self.d = {
+            "adcGain": 0,
+            "autoExposure": 0,
+            "cameraId": 0,
+            "captureId": 1,
+            "exposureTime": 111,
+            "imageDepth": 1,
+            "imageHeight": 2048,
+            "imageId": 0,
+            "imageWidth": 2048,
+            "immediateDownloadInfo": 10,
+            "instrument_name": "NavCam Left",
+            "lobt": 1700921056,
+            "offset": 0,
+            "outputImageMask": 2,
+            "outputImageType": "JBIG2_IMAGE",
+            "padding": 0,
+            "ppaGain": 0,
+            "processingInfo": 20,
+            "slog": False,
+            "stereo": 1,
+            "temperature": 0,
+            "voltageRamp": 0,
+        }
+
+    def test_no_image(self):
+        rp = cw.make_raw_product(self.d, None, None)
+        self.assertIsInstance(rp, RawProduct)
+
+    @patch("vipersci.vis.pds.create_raw.write_tiff")
+    @patch(
+        "vipersci.vis.pds.create_raw.tif_info",
+        return_value={
+            "file_byte_offset": 10,
+            "file_creation_datetime": datetime.fromtimestamp(1700921056, timezone.utc),
+            "file_data_type": "UnsignedMSB2",
+            "file_md5_checksum": "md5",
+            "file_path": "dummy.tif",
+            "lines": 2048,
+            "samples": 2048,
+        },
+    )
+    def test_image(self, mock_tif_info, mock_write_tiff):
+        image = np.array([[5, 5], [5, 5]], dtype=np.uint16)
+        rp = cw.make_raw_product(self.d, image, Path("outdir/"))
+        self.assertIsInstance(rp, RawProduct)
+
+        prp = cw.make_raw_product(self.d, Path("dummy.tif"), Path("outdir/"))
+        self.assertIsInstance(prp, RawProduct)
+
+        bad_d = self.d
+        bad_d["imageWidth"] = 2
         self.assertRaises(
-            IndexError,
-            pid.VIPERID,
-            datetime.date(2024, 1, 1),
-            datetime.time(1, 1, 1),
-            "aim",
-            "more than three arguments",
-        )
-        self.assertRaises(
-            IndexError,
-            pid.VIPERID,
+            ValueError, cw.make_raw_product, bad_d, image, Path("outdir/")
         )
 
-    def test_init_zero_args(self):
-        self.assertRaises(IndexError, pid.VIPERID)
 
-    def test_repr(self):
-        s = "This is a VIPER ID: 220117-010101-ncl"
-        p = pid.VIPERID(s)
-        self.assertEqual("VIPERID('220117-010101-ncl')", repr(p))
+class TestTIFF(unittest.TestCase):
+    @patch("vipersci.vis.pds.create_raw.datetime", wraps=datetime)
+    def test_tif_info(self, mock_datetime):
+        info = {
+            "ifds": [
+                {
+                    "tags": {
+                        273: {"data": (10,)},
+                        258: {"data": (16,)},
+                        256: {"data": (2048,)},
+                        257: {"data": (2048,)},
+                    }
+                },
+            ],
+            "bigEndian": "MSB",
+        }
+        mock_datetime.fromtimestamp.return_value = datetime.fromtimestamp(
+            1700921056, timezone.utc
+        )
 
-    def test_lt(self):
-        p1 = pid.VIPERID("231120-010101-acl")
-        p2 = pid.VIPERID("231121-010101-acl")
-        p3 = pid.VIPERID("231121-010102-acl")
-        p4 = pid.VIPERID("231121-010102-ncl")
-        self.assertTrue(p1 < p2)
-        self.assertTrue(p2 < p3)
-        self.assertTrue(p3 < p4)
+        mock_path = create_autospec(Path)
+        mock_path.name = "dummy.tif"
 
-    def test_datetime(self):
-        p = pid.VIPERID("231121-010101-acl")
-        self.assertEqual(p.datetime(), datetime.datetime(2023, 11, 21, 1, 1, 1))
+        with patch("vipersci.vis.pds.create_raw.open", mock_open(read_data=b"test")):
+            with patch("vipersci.vis.pds.create_raw.read_tiff", return_value=info):
+                d = cw.tif_info(mock_path)
+
+                self.assertEqual(d["file_byte_offset"], 10)
+                self.assertEqual(d["file_data_type"], "UnsignedMSB2")
+
+    @patch("vipersci.vis.pds.create_raw.imsave")
+    def test_write_tiff(self, mock_imsave):
+        pid = pds.VISID("231125-143859-ncl-a")
+        image = np.array([[5, 5], [5, 5]], dtype=np.uint16)
+        outpath = cw.write_tiff(pid, image, Path("dummy/"))
+        self.assertEqual(outpath, Path("dummy/231125-143859-ncl-a.tif"))
+        mock_imsave.assert_called_once_with(
+            str(outpath),
+            image,
+            check_contrast=False,
+            description="VIPER NavCam Left 231125-143859-ncl-a",
+            metadata=None,
+        )