+++ CHANGELOG.rst
@@ -28,6 +28,42 @@ and the release date, in year-month-day format (see examples below).
 Unreleased
 ----------
 
+Added
+^^^^^
+- pds.datetime.fromisozformat() function.
+- pds.pid.VISID.compression_class() function.
+- pds.xml.py added, very minimal, functionality may be moved.
+- vis.db.raw_products.RawProduct.from_xml() function.
+- vis.db.raw_products.RawProduct.asdict() function.
+- vis.pds.create_raw.check_bit_depth() function.
+- vis.db.create_vis_dbs convenience program to add empty tables to database.
+
+Changed
+^^^^^^^
+- Updated templates and modules for PDS information model 18.
+- vis.db.raw_products.RawProduct has some improved error-checking in __init__() and
+  validate_datetime_asutc().
+- vis.db.raw_products.RawProduct product_id column is now unique in database.
+- vis.db.raw_products.RawProduct md5_checksum changed to file_md4_checksum to
+  clearly associate it with the other properties that begin with "file_".
+- vis.pds.create_raw.tiff_info() no longer raises an error if a bit depth other than 16
+  is provided.
+- vis.pds.create_raw now creates .JSON output files by default instead of XML PDS4
+  labels, but XML files can still be made.
+
+
+Fixed
+^^^^^
+- carto.tri2gpkg.replace_with() now correctly returns a float value in all circumstances.
+- pds.pid.VIPERID.datetime() now properly returns datetimes with a UTC timezone.
+- vis.db.raw_products.RawProduct.label_dict() now correctly sets sample_bits and
+  sample_bit_mask if the image is a SLoG image.
+
+Removed
+^^^^^^^
+- Many modules still had if __name__ == "__main__" constructs from early development
+  which are now not needed with the entry points in setup.cfg.
+
 
 0.2.0 (2022-11-07)
 ------------------